Add query route to server
The `/plugins` route in the server queries all loaded plugins and will
return data from each of them.

All plugins are expected to provide a `/plugins/<pluginName>` route that
responds with all relevant data from that plugin.

Plugins are loaded at runtime based on the files in the /plugins folder.
It is assumed that the plugins folder only contains JS files, and each
one is meant to be loaded as a plugin.