Sort plugin data by priority
Plugin data on the `query` endpoint will be sorted by the priority of
the data being returned.

The documentation mentions that acceptable values for priority are
between 1 and 100, but currently no capping happens on priority values.
Plugins are assumed to be 'good actors' in the sense that they return
useful data-- this includes letting low-priority data sink to the bottom
of the pile and surfacing important information, which means reporting
reasonable priority numbers.