Implement back-in-time plugin
The back-in-time plugin checks for back-in-time backup directories in
the specified directory (environment variable) and will inform the user
how long it has been since the last successful backup.

If there has been a backup within the last 24 hours the backup situation
is treated as low-priority. If the last backup was more than 24 hours
ago the priority climbs as the number of days since the last backup
increases. Currently none of the priority values or backup thresholds
are configurable.