Add TTC Alert plugin skeleton
Basic files for TTC alert plugin. The TTC Alert will scrape the TTC
alerts website and scan through it for user-defined routes that are
travelled frequently or might interest the user. If there are any alerts
about routes that the user is interested in, the plugin will send back a
high-priority payload with the relevant alert attached.