Implement TTC Alert plugin behaviour
Basic TTC Alert functionality works. If there is an alert affecting the
routes a user takes regularly (usually commute times) the user will be
notified. There is currently no differentiation between important or
unimportant routes, nor is there a consideration for multiple alerts
affecting a user. Routes must also be entered with the correct spelling
and punctuation, or they will not be matched.