Implement disk usage plugin
Allows users to monitor their disks to see when a disk is getting full.
Disk usage warnings are capped at a priority of 90 (since full disks are
rarely life-threatening) and any disk that is less than 60% full is not
even registered as a minor problem.

Magic numbers of 2 and 85 are used for "no concern" and "high concern"
fairly arbitrarily; improper configuration should be surfaced, but is
not the end of the world (85) and if all disks are appropriately empty
then it's nice to see a friendly message to that effect (2).