Add umbrella alert priority logic
The decision on how important it is to have an umbrella is based on a
couple of factors. If there's a lot of rain in the next 3 hours, the
user almost certainly wants an umbrella (currently this results in a 50
priority message). If it is raining heavily for the next twelve hours,
the urgency increases.

It is also assumed that if it will rain in the near future then it is
less important for the user to know about rain after that-- they will
already need an umbrella. Additional rain later still increases the
urgency of the notification, but the change is not additive or linear.

Another assumption is that there is an amount of rain that _definitely_
requires an umbrella; here it has been set to 12mm in 3 hours, which may
not be a great estimate.

The final assumption is that rain now is more important than rain later.
Users should know about rain up to 12 hours in the future, but if it
will only start raining in 9 hours and the rest of the day will be dry
it is less likely that the user cares than if it will begin raining in
the next 3 hours.