Fix comparison in temperature difference
The logic is that when it's early in the morning (or any time before
evening, because I might be up late) I want to know how the coming day
compares to yesterday. Any time _after_ evening begins then I'm done
going outside and instead I'm looking to the next day to see what it
will be like.