Update TTC alert priorities
Elevator alerts should not be prioritized the same as actual line
closures (at least for those who can climb stairs/use escalators
freely). In the future it would be good to have some toggle for whether
or not elevator alerts are actually relevant to the user, but that
doesn't need to happen yet.

The terminology the TTC uses for its elevator alerts seems to have
changed (used to be "Elevator Alert: ..."), which may be worth keeping
an eye on.