Add update logic to nagbot
The Form API has to work with just POST and GET calls (due to the
limitations of the HTML Form spec), so a separate API has been made for
those calls. If the user has javascript enabled the form calls should be
intercepted by a JS handler and an AJAX call should be made instead of
the Form call, but for now it's all form calls all the time.